+++ Psicologia-Web/app/Http/Controllers/PlanosController.php
@@ -14,7 +14,7 @@ class PlanosController extends Controller
      */
     public function index()
     {
-        $planos = Planos::where('users_id', auth()->id())
+        $planos = Planos::where('user_id', auth()->id())
             ->orderBy('nome', 'asc')->get();
 
         return view("planos.index", ['planos' => $planos]);
@@ -28,6 +28,10 @@ class PlanosController extends Controller
      */
     public function create()
     {
+        if (auth()->user()->rules_id === 4) {
+            abort(403, 'Acesso nao autorizado.');
+        }
+
         return view('planos.create');
     }
 
@@ -39,12 +43,16 @@ class PlanosController extends Controller
      */
     public function store(Request $request)
     {
+        if (auth()->user()->rules_id === 4) {
+            abort(403, 'Acesso nao autorizado.');
+        }
+
         $data = $request->all();
-        $data['users_id'] = auth()->id();
+        $data['user_id'] = auth()->id();
 
         Planos::create($data);
 
-        $planos = Planos::where('users_id', auth()->id())
+        $planos = Planos::where('user_id', auth()->id())
             ->orderBy('nome', 'asc')->get();
 
         return redirect()->route("planos.index", compact('planos'));
@@ -56,9 +64,13 @@ class PlanosController extends Controller
      * @param  \App\Models\Planos  $planos
      * @return \Illuminate\Http\Response
      */
-    public function show(Planos $planos)
+    public function show(Planos $plano)
     {
-        //
+        if ($plano->user_id !== auth()->id()) {
+            abort(403, 'Acesso não autorizado.');
+        }
+
+        return view('planos.show', compact('plano'));
     }
 
     /**
@@ -67,9 +79,17 @@ class PlanosController extends Controller
      * @param  \App\Models\Planos  $planos
      * @return \Illuminate\Http\Response
      */
-    public function edit(Planos $planos)
+    public function edit(Planos $plano)
     {
-        //
+        if (auth()->user()->rules_id === 4) {
+            abort(403, 'Acesso nao autorizado.');
+        }
+
+        if ($plano->user_id !== auth()->id()) {
+            abort(403, 'Acesso não autorizado.');
+        }
+
+        return view('planos.edit', compact('plano'));
     }
 
     /**
@@ -79,9 +99,20 @@ class PlanosController extends Controller
      * @param  \App\Models\Planos  $planos
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Planos $planos)
+    public function update(Request $request, Planos $plano)
     {
-        //
+        if (auth()->user()->rules_id === 4) {
+            abort(403, 'Acesso nao autorizado.');
+        }
+
+        if ($plano->user_id !== auth()->id()) {
+            abort(403, 'Acesso não autorizado.');
+        }
+
+        $data = $request->all();
+        $plano->update($data);
+
+        return redirect()->route('planos.index');
     }
 
     /**
@@ -92,9 +123,13 @@ class PlanosController extends Controller
      */
     public function destroy(Planos $plano)
     {
+        if (auth()->user()->rules_id === 4) {
+            abort(403, 'Acesso nao autorizado.');
+        }
+
         $plano->delete();
 
-        $planos = Planos::where('users_id', auth()->id())->orderBy('nome', 'asc')->get();
+        $planos = Planos::where('user_id', auth()->id())->orderBy('nome', 'asc')->get();
 
         return redirect()->route("planos.index", compact('planos'));
     }
